viewmodel unit test added

+++ app/src/main/java/com/vsantander/vehicleschallenge/data/remote/model/CoordinateTO.kt
@@ -4,8 +4,8 @@ import com.google.gson.annotations.SerializedName
 
 data class CoordinateTO (
         @SerializedName("latitude")
-        val latitude: Float,
+        val latitude: Double,
 
         @SerializedName("longitude")
-        val longitude: Float
+        val longitude: Double
 )
